More Xcode recommended updates
Some crashes in the stats screens seem to have vanished

+++ UniversalClient/Shared/Models/Archaeology.m
@@ -42,7 +42,7 @@
 
 
 - (NSString *)description {
-	return [NSString stringWithFormat:@"secondsRemaining:%i, glyphs:%@", self.secondsRemaining, self.glyphs];
+	return [NSString stringWithFormat:@"secondsRemaining:%li, glyphs:%@", (long)self.secondsRemaining, self.glyphs];
 }
